migration has been done

+++ app/src/main/java/com/example/myapplication/roomdb/RoomDBAdapter.java
@@ -1,16 +1,17 @@
 package com.example.myapplication.roomdb;
 
+import android.app.AlertDialog;
 import android.content.Context;
 import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.PopupMenu;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.widget.AppCompatImageButton;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.room.Room;
 
 import com.example.myapplication.R;
 
@@ -47,28 +48,49 @@ public class RoomDBAdapter extends RecyclerView.Adapter<RoomDBAdapter.viewHolder
         holder.tvStudentAddress.setText(model.getsAddress());
         holder.tvStudentPhoneNo.setText(model.getsMobileNo());
         holder.tvStudentQualified.setText(model.getsQualification());
+        holder.tvGender.setText(model.getsGender());
 
-        holder.ivBtnDeleteRowData.setOnClickListener(v -> {
-            StudentDataBase dataBase = Room.databaseBuilder(context, StudentDataBase.class,
-                    "student_database").allowMainThreadQueries().build();
-
-            StudentDao studentDao = dataBase.getStudentDao();
-            studentDao.deletedStudentData(model.getsId());
-            studentsModels.remove(position);
-            notifyDataSetChanged();
-
+        holder.ivOptionMenu.setOnClickListener(v -> {
+            PopupMenu popupMenu = new PopupMenu(context, holder.ivOptionMenu);
+            popupMenu.inflate(R.menu.option_menu_list);
+            popupMenu.setOnMenuItemClickListener(item -> {
+                int itemId = item.getItemId();
+                if (itemId == R.id.menu_update) {
+                    Intent intent = new Intent(context, UpdateStudentRDBActivity.class);
+                    intent.putExtra(STUDENT_ID, String.valueOf(model.getsId()));
+                    intent.putExtra(STUDENT_NAME, model.getsName());
+                    intent.putExtra(STUDENT_EMAIL, model.getsEmail());
+                    intent.putExtra(STUDENT_ADDRESS, model.getsAddress());
+                    intent.putExtra(STUDENT_PHONE_NO, model.getsMobileNo());
+                    intent.putExtra(STUDENT_QUALIFICATION, model.getsQualification());
+                    context.startActivity(intent);
+                } else if (itemId == R.id.menu_delete) {
+                    new AlertDialog.Builder(context)
+                            .setTitle("Delete")
+                            .setMessage("if want to delete this record")
+                            .setCancelable(false)
+                            .setPositiveButton("yes", (dialog, which) -> {
+                                /*StudentDataBase dataBase = Room.databaseBuilder(context, StudentDataBase.class,
+                                        "student_database")
+                                        .allowMainThreadQueries()
+                                        .build();*/
+                                StudentDataBase dataBase = StudentDataBase.getInstance(context);
+                                StudentDao studentDao = dataBase.getStudentDao();
+                                studentDao.deletedStudentData(model.getsId());
+                                studentsModels.remove(position);
+                                notifyDataSetChanged();
+                            })
+                            .setNegativeButton("no", (dialog, which) -> {
+                                dialog.dismiss();
+                            })
+                            .create()
+                            .show();
+                }
+                return false;
+            });
+            popupMenu.show();
         });
 
-        holder.itemView.setOnClickListener(v -> {
-            Intent intent = new Intent(context, UpdateStudentRDBActivity.class);
-            intent.putExtra(STUDENT_ID, String.valueOf(model.getsId()));
-            intent.putExtra(STUDENT_NAME, model.getsName());
-            intent.putExtra(STUDENT_EMAIL, model.getsEmail());
-            intent.putExtra(STUDENT_ADDRESS, model.getsAddress());
-            intent.putExtra(STUDENT_PHONE_NO, model.getsMobileNo());
-            intent.putExtra(STUDENT_QUALIFICATION, model.getsQualification());
-            context.startActivity(intent);
-        });
     }
 
     @Override
@@ -82,7 +104,8 @@ public class RoomDBAdapter extends RecyclerView.Adapter<RoomDBAdapter.viewHolder
         private TextView tvStudentAddress;
         private TextView tvStudentPhoneNo;
         private TextView tvStudentQualified;
-        private AppCompatImageButton ivBtnDeleteRowData;
+        private TextView tvGender;
+        private ImageView ivOptionMenu;
 
         public viewHolder(@NonNull View itemView) {
             super(itemView);
@@ -91,7 +114,8 @@ public class RoomDBAdapter extends RecyclerView.Adapter<RoomDBAdapter.viewHolder
             tvStudentAddress = itemView.findViewById(R.id.tv_address);
             tvStudentPhoneNo = itemView.findViewById(R.id.tv_phone_no);
             tvStudentQualified = itemView.findViewById(R.id.tv_qualified);
-            ivBtnDeleteRowData = itemView.findViewById(R.id.iv_btn_delete);
+            tvGender = itemView.findViewById(R.id.tv_gender);
+            ivOptionMenu = itemView.findViewById(R.id.option_menu);
         }
     }
 }
